src/test/kotlin/y24/Day23Test.kt
package y24

import common.puzzle.Input
import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.Test

internal class Day23Test {
    private val sample = Input("""
    kh-tc
    qp-kh
    de-cg
    ka-co
    yn-aq
    qp-ub
    cg-tb
    vc-aq
    tb-ka
    wh-tc
    yn-cg
    kh-ub
    ta-co
    de-co
    tc-td
    tb-wq
    wh-td
    ta-ka
    td-qp
    aq-cg
    wq-ub
    ub-vc
    de-ta
    wq-aq
    wq-vc
    wh-yn
    ka-de
    kh-ta
    co-tc
    wh-qp
    tb-vc
    td-yn
    """.trimIndent())

    private val day = Day23(sample)

    @Test
    fun solveLevel1() {
        assertEquals(7, day.solveLevel1())
    }

    @Test
    fun solveLevel2() {
         assertEquals("co,de,ka,ta", day.solveLevel2())
    }
}
